внесены изменения в программу 1

+++ PrG1/Program.cs
@@ -1,11 +1,11 @@
 ﻿Console.WriteLine("Введите 3 числа:");
 Console.WriteLine("Введите первое число:");
-int a = int .Parse(Console.ReadLine());
+double a = int.Parse(Console.ReadLine());
 Console.WriteLine("Введите второе число:");
-int  b = int .Parse(Console.ReadLine());
+double  b = int.Parse(Console.ReadLine());
 Console.WriteLine("Введите третье число:");
-int  c = int .Parse(Console.ReadLine());
+double  c = int.Parse(Console.ReadLine());
 
-int sum = a + b + c;
- double arg = sum/3;
+double sum = a + b + c;
+double arg = sum/3;
 Console.WriteLine($"Среднее арифметическое равно {arg} ");
